Works in progress for supporting specifications.
The 0.2 workspec schema is proposed as a mechanism for
conveying the record of work requested by the client.

gmxapi is an evolving framework that is currently a full
stack implementation, but which can hopefully be
decomposed in future releases into a set of adapters
between GROMACS tools, workflow and workload middleware,
and user interface client support.

Refs #8