Clarify workflow loop phases.
There was an implicit conversion of "conformation" to SimulationInput,
which is confusing and likely unfeasible.

The options are either
(a) to provide two parameters to `simulation_iteration`, or
(b) to explicitly identify input modification as a separate phase of the loop.

Both versions of simulation_iteration are demonstrated.
The simulation_and_analysis_loop body has been updated
to reflect option b.