Convert pytest fixtures to use command line options.
Add `--rp-resource` `--rp-access` and `--rp-venv` command line options
that can be appended to a pytest command line with access to conftest.py

The resource and access mode need to match a resource definition
available to the invoking user.

The path provided to --rp-venv is not yet used.

Hopefully, this makes it easier to run the same tests with
different target resources and execution modes.

The travis-ci configuration has been expanded to try to
exercise more execution modes.