Add some support for a debug server.
Assume that tests will be launched manually from a shell
after activating an appropriate venv.

It may be insufficient for pytest to amend
the PATH in order to meet the assumptions of
RP forked process environments.
Remove the code for this from conftest.py to improve clarity.

Add a command line option to enable the PyCharm version of pydevd.
(Assumes a fairly specific debug server configuration until someone
needs it to be more flexible.)