Update test for RP future.
Give the wrapper task a chance to start
before sending the cancellation.

Allow that the wrapped RP task may complete
before the cancellation signal can be sent by the wrapper.

Note that, unless paused by an attached debugger,
the RP test task should not have a chance to complete, but it may in the
 general case.