Always stage scheduler config.
The only current mechanism for passing information
to the Raptor Master task is with a command line argument,
which is interpreted as the path to a JSON file
containing a serialized ru.Config.

The file is now generated by the client
(in terms of helper classes in scalems.radical.scalems_rp_agent)
and staged with the Master task submission.