Add a workaround to a test.
Some resources are not released fast enough sometimes while running the
full tests. Inserting a ten-second sleep in a test that seems to create
a Pilot at just the wrong time seems to be an effective workaround that
will make testing easier in the near term. Hopefully this isn't
permanent.