Remove extra venv cache.
The `setup-python@v4` GitHub Action seems to break our venv cache
(based on the `cache@v3` Action), but has its own `pip` cache.

For the moment, I'm removing our cache of the built `testenv` venv.
This means that we will rebuild LAMMPS on every test pipeline. We can
try to resolve that wastage with `ccache` in a follow-up change.