remove dead code
I can't remember where this came from. In at least one version of Python, there was some sort of internal metaprogramming (maybe isinstance() or functools.singledispatch? or maybe a check related to ABC or metaclass hierarchy) that spit an error at some point when a concrete class did not find a `__init__()` before finding a typing.Protocol when traversing the MRO. It doesn't seem to be causing a problem here and now.

If we encounter the problem again, we should document it better.